moved impl of save_to_file() function to serial namespace
